Begin moving common reducers/components/actions/constants to shared libraries module
