+++ client/app/bundles/AdminView/components/HomePage/index.jsx
@@ -2,7 +2,7 @@
 import React, { Component } from "react";
 import PropTypes from "prop-types";
 import { connect } from "react-redux";
-import * as pageNavActions from "../../actions/pageNavActions";
+import * as pageNavActions from "libs/actions/pageNavActions";
 
 class HomePage extends Component {
   static propTypes = {
