+++ client/app/bundles/AdminView/reducers/index.jsx
@@ -1,8 +1,12 @@
 import { combineReducers } from "redux";
 import { routerReducer } from "react-router-redux";
-import userReducer, { initialState as userState } from "./userReducer";
+import userReducer, {
+  initialState as userState,
+} from "libs/reducers/userReducer";
+import pageNavReducer, {
+  initialState as pageNavState,
+} from "libs/reducers/pageNavReducer";
 import usersReducer, { initialState as usersState } from "./usersReducer";
-import pageNavReducer, { initialState as pageNavState } from "./pageNavReducer";
 
 export const initialStates = {
   user: userState,
