+++ client/app/libs/constants/pageNavConstants.js
@@ -1,0 +1,5 @@
+export const PAGE_HAS_NAVIGATED = "PAGE_HAS_NAVIGATED";
+
+export default {
+  PAGE_HAS_NAVIGATED,
+};
