+++ client/app/libs/constants/userConstants.js
@@ -1,0 +1,11 @@
+export const FETCH_USER_SUCCESS = "FETCH_USER_SUCCESS";
+export const FETCH_USER_FAILURE = "FETCH_USER_FAILURE";
+export const SET_IS_FETCHING = "SET_IS_FETCHING";
+export const UPDATE_USER_SUCCESS = "UPDATE_USER_SUCCESS";
+export const UPDATE_USER_FAILURE = "UPDATE_USER_FAILURE";
+export const SET_IS_UPDATING = "SET_IS_UPDATING";
+export const SET_UPDATE_SUCCESS = "SET_UPDATE_SUCCESS";
+export const CHANGE_SETTING_VALUES = "CHANGE_SETTING_VALUES";
+
+export const USER_PATH = "/users";
+export const USER_ME_PATH = `${USER_PATH}/me`;
