+++ client/app/libs/reducers/pageNavReducer.js
